stm: adds patches for building kernel 6.6.y as WIP, update README
this patch is for adding a supported kernel
release as WIP and adds to a dir marked as WIP
to highlight it:
stm_iio_patches/6.6.y.WIP

merging the drivers to the WIP marked (6.6.y) kernel
release could result in a failing build.

This patch applies to kernel.org mainline branch
linux-6.6.y

Readme.md file is accordingly updated to reflect current
status.

Change-Id: Ie162471e25e7f193f5287563603b320875f160de
Signed-off-by: Matteo Dameno <matteo.damenom@st.com>