drivers:iio:stm:imu:st_lsm6dsvx: add support of IIO software trigger
Added the possibility to use the IIO buffer interface
even when a hardware interrupt line is not available.
In this case it will be necessary to connect an external
software trigger via the iio-trig-hrtimer module.

NOTE: using the software triggered mode the hw FIFO of the
sensor is excluded and therefore the functionality of the
driver is limited because it is no longer possible to
read data from the sensors which are read via the FIFO.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I946894554db40e7f42f0f2dd9ce2aa3ca839765e