drivers:iio:stm:imu:st_lsm6dsvx: calculate number of sample to be discarded when polling axis raw data
When reading the accel and gyro axis raws data, a certain
number of samples must be discarded after turning on the
sensor, furthermore a minimum waiting time between the
power off and the subsequent power on of the sensor should
be guaranteed.
This patch introduces the calculation of these times as
described in AN5763.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ia2ac815347ac45500b7ea2e399b7fb892925456a