drivers:iio:stm:imu:st_lsm6dsrx: porting timesync patchsets from asm330lhhx
This patch is a porting of the following two patches:

Commit 12a114731904 ("drivers:iio:stm:imu:st_asm330lhhx: rework timesync to guarantee a minimum number of syncs for all sensors")

With this patch we guarantee at least 10 fast sync samples at
each timesync instance to have a good estimate of the sensor
timestamp.
In the past the number of fast samples was not guaranteed to
each sensor so it could happen that the fast sync samples ran
out too soon before the sensor even started sending data.

Commit a6efac2105ba ("drivers:iio:stm:imu:st_asm330lhhx: channel optimization when timesync is disabled")

When the timesync function is disabled the IIO_COUNT type channels
used for sensor timestamp synchronization management are not used
and can be removed. This change also impacts the FIFO reading
function since the copy of the hardware timestamp on the related
channel is no longer required.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I21999e7e1cf4cec3fe6ec94f4dd33d5a50da1793