drivers:iio:stm:imu:st_lsm6dsv16bx: rework self-test procedure
The self test procedure for the accelerometer did not handle
the XL_ST_OFFSET bit, also the x, y, and z output registers
were not consistent with the data sheet.
In this patch also the functions for the self-test management
between accel and gyro have been put in common in order to
reduce redundant code.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I7850796afe44abeef3145529c3889f20964c9d4a