drivers:iio:stm:magnetometer:st_mag40: fix github issue "Driver crash in ISR for LSM303AGR Magnetometer ( st_mag40* ) #28"
It has been noticed that if the device is not reset during a reboot it could
generate an interrupt that if triggered before the devm_request_threaded_irq
call could invoke the st_mag40_trigger_thread_handler function without
having initialized the cdata->iio_trig structure.
This patch fixes the issue and also introduces the sensor reset both during
the driver removal phase and during the module probe phase.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ib9a2f5d89d18502a80730ac76340ab0f3be73b38