drivers:iio:stm:imu:st_asm330lhhx: remove call to enable sensor function during self-test procedure
During the sensor's self-test procedure, the call to the
enable function incorrectly sets the ODR to an incorrect
value. To follow the correct procedure, it is necessary
to save the sensor configuration and set the ODR directly
in the configuration parameters.
Afterwards, the configuration values will be restored to
their correct values prior to the self-test procedure.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I04d0a2cad17e3bf9926cd09f77927a11958fad04