drivers:iio:stm:accel:st_lis2ds12: Reworked read raw data samples procedure
The interrupt is incorrectly enabled in the raw data read function,
which could cause the handler to be triggered before the timeout
expires.

In this commit, a function has now been added to enable accel, which
does not enable the interrupt line. It also waits for a time that
depends on the ODR configured on the sensor.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ic6d9ebb5d80970828396e53ae5b7862d4ca494e4