drivers:iio:stm:imu:st_asm330lhhx: add support to WAKE_THS_W
When setting wake-up detection thresholds below one LSB,
there is a risk that the threshold register will be set
to zero, causing continuous wake-up events.

By using the WAKE_THS_W bit, it is possible to use an LSB
that is four times smaller, allowing for thresholds lower
than the default value expected when WAKE_THS_W is set.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ia8df61879bb551c2d5c28c8f98d9dd56eda41bc5